==> 2-EDA-spline-fitting-in_counts-out_counts-exponential-points-and-peaks-on-weight-timestamps.Rmd <==
---
title: "Spline Fitting All Hives Individually Over 1 Day, Simulating Missing Data for Peaks & Points of Exponentiation on in_count & out_count"
author: "April Ainsworth"
date: "2024-06-14"
output: pdf_document
---

```{r echo=FALSE}
library(splines)
library(dplyr)
library(ggplot2)
library(RColorBrewer)
library(lubridate)
library(pracma)

all_hives <- read.csv("C:/ncf-graduate-school/internship-USDA/project8-almond-pollination/data/final_merged_data.csv")
# print(str(all_hives))
sun_times <- read.csv("C:/ncf-graduate-school/internship-USDA/project8-almond-pollination/data/sun_times.csv")

# hive 12 has most observations of all hives; 2/20 is a good day to work with (no inspections, no odd weather)

one_day = "2023-02-20"
sunrise = sun_times %>% filter(date == one_day) %>% select(sunrise)
sunrise_posix <- ymd_hms(sunrise, tz = "PST8PDT")
# sunrise_posix <- force_tz(sunrise_posix, tzone = "PST8PDT")
# sunrise_posix <- parse_date_time(sunrise, orders = "Ymd HMS OSz", tz = "UTC")
# print(sunrise)
# print(sunrise_posix)

all_hives$weight_timestamp <- as.POSIXct(all_hives$weight_timestamp, format = "%Y-%m-%d %H:%M:%S")
all_hives$weight_timestamp_numeric <- as.numeric(all_hives$weight_timestamp)

start_time <- as.POSIXct("2023-02-20 05:00:00", format = "%Y-%m-%d %H:%M:%S")
# start_time <- as.POSIXct(sunrise_posix, format = "%Y-%m-%d %H:%M:%S")
midnight <- as.POSIXct("2023-02-20 23:59:59", format = "%Y-%m-%d %H:%M:%S")

hive_series = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 23, 24)

for (one_hive in hive_series) {  
  # print(paste0("Length of Weight Timestamps for Hive ", one_hive))
  # print(length(unique(one_hive_one_day$weight_timestamp_numeric)))

  one_hive_one_day <- all_hives %>% filter(hive_id_anon == one_hive, weight_timestamp_numeric > start_time & weight_timestamp_numeric <= midnight)
  # print(one_hive_one_day)
  
  # sort by timestamp, create diff columns, merge timestamps with coalesce()
  one_hive_one_day <- one_hive_one_day %>% arrange(weight_timestamp_numeric) %>% mutate(in_count_diff = coalesce(in_count - lag(in_count), 0), out_count_diff = coalesce(out_count - lag(out_count), 0))
  
  # print(one_hive_one_day$in_count_diff)
  # print(paste0("2nd Marker Length of Numeric Timestamps for Hive ", one_hive))
  # print(length(unique(one_hive_one_day$weight_timestamp_numeric)))
  
  spline_fit <- smooth.spline(one_hive_one_day$weight_timestamp_numeric, one_hive_one_day$in_count_diff, spar = 0.5)
  spline_fit_out_count <- smooth.spline(one_hive_one_day$weight_timestamp_numeric, one_hive_one_day$out_count_diff, spar = 0.5)
    
#  plot(one_hive_one_day$weight_timestamp_numeric, one_hive_one_day$in_count_diff, main = "Smooth Spline with Subset Data", xlab = #  #"weight_timestamp_numeric", ylab = "in_count_diff", pch = 19, col = "blue")
#  lines(spline_fit, col = "red", lwd = 2)
  
  spar_vals <- c(0.1, 0.4, 0.6)
  
  tst <- lapply(spar_vals, FUN = function(i, dat) {
    smooth.spline(dat$weight_timestamp_numeric, dat$in_count_diff, spar = i)
    smooth.spline(dat$weight_timestamp_numeric, dat$out_count_diff, spar = i)
  },
  dat = one_hive_one_day 
  )
  
  plot_title_one_hive_one_day <- paste("in_count vs weight timestamps: Smooth Spline with Spar = ", paste(spar_vals, collapse = ", "))
  plot_title_one_hive_one_day_out_count <- paste("out_count vs weight timestamps: Smooth Spline with Spar = ", paste(spar_vals, collapse = ", "))
  
  plot(one_hive_one_day$weight_timestamp, one_hive_one_day$in_count_diff, main = plot_title_one_hive_one_day, xlab = "weight_timestamp", ylab = "in_count_diff", pch = 16)
  lapply(tst, function(fit) lines(fit, lty = 3))
  
  plot(one_hive_one_day$weight_timestamp, one_hive_one_day$out_count_diff, main = plot_title_one_hive_one_day_out_count, xlab = "weight_timestamp", ylab = "in_count_diff", pch = 16)
  lapply(tst, function(fit) lines(fit, lty = 3))
  
  # simulate missingness based on peaks and points of exponentiation
  
  # find where exponentiation begins/ends
  calculate_derivatives <- function(smoothed_data) {
    first_derivative <- diff(smoothed_data) / diff(one_hive_one_day$weight_timestamp_numeric)
    second_derivative <- diff(first_derivative) / diff(one_hive_one_day$weight_timestamp_numeric[-1])
  
    # align lengths
    first_derivative <- c(NA, first_derivative)
    second_derivative <- c(NA, NA, second_derivative)
  
    list(first_derivative = first_derivative, second_derivative = second_derivative)
  }

  plots_exponential <- lapply(spar_vals, function(spar, data) {
    spline_fit <- smooth.spline(data$weight_timestamp_numeric, data$out_count_diff, spar = spar)
    smoothed_out_count <- predict(spline_fit, data$weight_timestamp_numeric)$y
    print(paste0("smoothed_out_count: ", smoothed_out_count))
    derivatives <- calculate_derivatives(smoothed_out_count)
  
    derivatives_df <- data %>%
      mutate(smoothed_out_count = smoothed_out_count,
           first_derivative = derivatives$first_derivative,
           second_derivative = derivatives$second_derivative)
  
  # find where exponential growth begins (significant positive second derivative)
  better_threshold <- mean(derivatives_df$second_derivative, na.rm = TRUE) + 2 * sd(derivatives_df$second_derivative, na.rm = TRUE)
  # likely need to use better_threshold because it'll adapt to varying colony sizes
  print(paste0("better threshold is: ", better_threshold))
  # threshold <- 0.01
  growth_start_time <- derivatives_df %>%
    filter(second_derivative > better_threshold) %>%
    slice(1) %>%
    pull(weight_timestamp)

  plot_growth <- ggplot(data, aes(x = weight_timestamp, y = out_count_diff)) +
    geom_line(color = "blue") +
    geom_line(aes(y = smoothed_out_count), color = "red") +
    geom_vline(xintercept = as.numeric(growth_start_time), linetype = "dashed", color = "green") +
    annotate("text", x = growth_start_time, y = max(data$out_count_diff, na.rm = TRUE) * 0.95, label = "Exponential Growth", color = "green", angle = 90, vjust = -0.5, hjust = 1) +
    labs(title = paste0("Exponential Growth of Bees Exiting the Hive (spar = ", spar, ")"),
         x = "Timestamp",
         y = "out_count_diff") +
    theme_minimal()
  
  return(plot_growth)
  }, data = one_hive_one_day)

  for (plot in plots_exponential) {
    print(plot)
  }

  
  time_gaps <- c(1, 2, 3, 4, 5, 6, 7, 8)
  # missing_start <- as.POSIXct("2023-02-20 10:00:00", format = "%Y-%m-%d %H:%M:%S")
  # missing_start <- as.POSIXct(sunrise_posix, format = "%Y-%m-%d %H:%M:%S")
  missing_start <- start_time
  missing_end_times <- missing_start + hours(time_gaps)
  print(missing_end_times)
  
  timestamps_remaining <- length(unique(one_hive_one_day$weight_timestamp_numeric)) # start with all timestamps available, stop when < 5
  print(paste0("Starting With ", timestamps_remaining, " Timestamps For Hive ", one_hive))
  # compare time gaps
  
  for (end_time in missing_end_times) {
    # print(paste0("end_time is now ", end_time)) # verified that values are correct per iteration
    end_time_formatted <- format(as.POSIXct(end_time, origin="1970-01-01"), "%Y-%m-%d %H:%M:%S")
    end_time <- as.POSIXct(end_time, origin = "1970-01-01")
    # end_time_formatted <- ymd_hms(end_time, tz = "PST8PDT") # this version outputs NA
    # if(length(unique(one_hive_one_day$weight_timestamp_numeric)) >= 5) {
    # timestamps_remaining <- timestamps_remaining - length(unique(one_hive_one_day$weight_timestamp_numeric))
      ### the number of timestamps that exists between start_time and end_time
    num_timestamps_in_gap <- sum(one_hive_one_day$weight_timestamp_numeric >= start_time & one_hive_one_day$weight_timestamp_numeric < end_time)
    timestamps_remaining <- timestamps_remaining - num_timestamps_in_gap
    # print(paste0("Timestamps Remaining Are: ", timestamps_remaining))
    if (timestamps_remaining >= 4) {
      # print("inside if/else block")
      one_hive_one_day$in_count_diff[one_hive_one_day$weight_timestamp >= missing_start & one_hive_one_day$weight_timestamp <= end_time] <- NA
      one_hive_one_day$out_count_diff[one_hive_one_day$weight_timestamp >= missing_start & one_hive_one_day$weight_timestamp <= end_time] <- NA
    
      data_without_na <- one_hive_one_day %>% filter(!is.na(in_count_diff))
      data_without_na_out_count <- one_hive_one_day %>% filter(!is.na(out_count_diff))
    
      # fit smooth.spline to completely available data (no NA values)
      spline_fit <- smooth.spline(data_without_na$weight_timestamp_numeric, data_without_na$in_count_diff)
      spline_fit_out_count <- smooth.spline(data_without_na$weight_timestamp_numeric, data_without_na$out_count_diff)
    
      # predict missing values with spline_fit
      missing_indices <- which(is.na(one_hive_one_day$in_count_diff))
      predicted_values <- predict(spline_fit, one_hive_one_day$weight_timestamp_numeric[missing_indices])$y
      
      missing_indices_out_count <- which(is.na(one_hive_one_day$out_count_diff))
      predicted_values_out_count <- predict(spline_fit_out_count, one_hive_one_day$weight_timestamp_numeric[missing_indices])$y
      # print(predicted_values)
    
      # impute the missing values
      one_hive_one_day$in_count_diff_imputed <- one_hive_one_day$in_count_diff
      one_hive_one_day$in_count_diff_imputed[missing_indices] <- predicted_values
      # print(one_hive_one_day$in_count_diff_imputed)
      one_hive_one_day$out_count_diff_imputed <- one_hive_one_day$out_count_diff
      one_hive_one_day$out_count_diff_imputed[missing_indices_out_count] <- predicted_values_out_count
    
      spline_df <- data.frame(
        weight_timestamp_numeric = one_hive_one_day$weight_timestamp_numeric,
        in_count_diff_spline = predict(spline_fit, one_hive_one_day$weight_timestamp_numeric)$y
      )
      # to get timestamps to appear human-readable later in plots
      spline_df <- spline_df %>% mutate(weight_timestamp = as.POSIXct(weight_timestamp_numeric, origin = "1970-01-01"))
      
      spline_df_out_count <- data.frame(
        weight_timestamp_numeric = one_hive_one_day$weight_timestamp_numeric,
        out_count_diff_spline = predict(spline_fit_out_count, one_hive_one_day$weight_timestamp_numeric)$y
      )
      
      spline_df_out_count <- spline_df_out_count %>% mutate(weight_timestamp = as.POSIXct(weight_timestamp_numeric, origin = "1970-01-01"))
      
      # print(end_time)
      
      # plot original and imputed data: in_count
      plot1 <- ggplot(one_hive_one_day, aes(x = weight_timestamp)) + 
        geom_point(aes(y = in_count_diff_imputed, color = "Imputed Data"), na.rm = TRUE) +
        geom_point(aes(y = in_count_diff, color = "Original Data"), na.rm = TRUE) +
        geom_line(data = spline_df, aes(x = weight_timestamp, y = in_count_diff_spline, color = "Spline Fit")) +
        geom_vline(xintercept = end_time, linetype = "dashed", color = "red") +
        annotate("text", x = end_time, y = max(one_hive_one_day$in_count_diff, na.rm = TRUE), label = "End Missing Data", color = "red", angle = 90, vjust = -0.5, hjust = 1) +
        labs(title = paste0("Hive ", one_hive, ": predict() Imputation of in_counts Missing Data Ending ", end_time_formatted), x = "Timestamp", y = "in_count_diff") +
        # scale_x_datetime(date_labels = "%Y-%m-%d %H:%M:%S") +
        scale_color_brewer(palette = "Set1") +
        theme_minimal() +
        theme(legend.title = element_blank())
      
      print(plot1)
    
      # make multiple splines: in_count
      splines_list <- lapply(spar_vals, function(spar) {
        smooth.spline(data_without_na$weight_timestamp_numeric, data_without_na$in_count_diff, spar = spar)
      })
    
      spline_dfs <- lapply(seq_along(splines_list), function(i) {
        data.frame(
          weight_timestamp_numeric = one_hive_one_day$weight_timestamp_numeric,
          in_count_diff_spline = predict(splines_list[[i]], one_hive_one_day$weight_timestamp_numeric)$y, 
          spar = paste0("in_count Spline Fit spar=", spar_vals[i])
        ) %>% mutate(weight_timestamp = as.POSIXct(weight_timestamp_numeric, origin = "1970-01-01"))
      })
      # spline_dfs <- spline_dfs %>% mutate(weight_timestamp = as.POSIXct(weight_timestamp_numeric, origin = "1970-01-01"))
    
      combined_spline_df <- do.call(rbind, spline_dfs)
    
      # Compare imputed data with actual data (before removal): in_count
      # Keep a copy of the original data before simulating missing data
      one_hive_one_day$in_count_diff_original <- one_hive_one_day$in_count - lag(one_hive_one_day$in_count)
      one_hive_one_day$in_count_diff_original[is.na(one_hive_one_day$in_count_diff_original)] <- 0
    
      plot2 <- ggplot(one_hive_one_day, aes(x = weight_timestamp)) +
        geom_point(aes(y = in_count_diff_imputed, color = "Imputed Data"), na.rm = TRUE) +
        geom_point(aes(y = in_count_diff_original, color = "Original Data"), na.rm = TRUE) +
        geom_line(data = combined_spline_df, aes(x = weight_timestamp, y = in_count_diff_spline, color = spar), size = 1) +
        geom_vline(xintercept = end_time, linetype = "dashed", color = "red") +
        annotate("text", x = end_time, y = max(one_hive_one_day$in_count_diff, na.rm = TRUE), label = "End Missing Data", color = "red", angle = 90, vjust = -0.5, hjust = 1) +
        # TODO: try this for timestamp aes(x = as.POSIXct(weight_timestamp_numeric, origin="1970-01-01")
        # geom_line(data = spline_df, aes(y = in_count_diff_spline, color = "Spline Fit")) + # change out this for combined_spline_df
        labs(title = paste0("Hive ", one_hive, ": in_count Imputations For Timespan Ending ", end_time), x = "Timestamp", y = "in_count_diff") +
        # scale_x_datetime(date_labels = "%Y-%m-%d %H:%M:%S") +
        scale_color_brewer(palette = "Set1") +
        theme_minimal() +
        theme(legend.title = element_blank())
      
      print(plot2)
      
      # plot original and imputed data: out_count
      plot3 <- ggplot(one_hive_one_day, aes(x = weight_timestamp)) +
        geom_point(aes(y = out_count_diff_imputed, color = "Imputed Data"), na.rm = TRUE) +
        geom_point(aes(y = out_count_diff, color = "Original Data"), na.rm = TRUE) +
        geom_line(data = spline_df_out_count, aes(x = weight_timestamp, y = out_count_diff_spline, color = "Spline Fit")) +
        geom_vline(xintercept = end_time, linetype = "dashed", color = "red") +
        annotate("text", x = end_time, y = max(one_hive_one_day$in_count_diff, na.rm = TRUE), label = "End Missing Data", color = "red", angle = 90, vjust = -0.5, hjust = 1) +
        labs(title = paste0("Hive ", one_hive, ": predict() Imputation of out_counts Missing Data, Ending ", end_time), x = "Timestamp", y = "out_count_diff") +
        # scale_x_datetime(date_labels = "%Y-%m-%d %H:%M:%S") +
        scale_color_brewer(palette = "Set1") +
        theme_minimal() +
        theme(legend.title = element_blank())
      
      print(plot3)
    
      # make multiple splines: in_count
      splines_list_out_count <- lapply(spar_vals, function(spar) {
        smooth.spline(data_without_na_out_count$weight_timestamp_numeric, data_without_na_out_count$out_count_diff, spar = spar)
      })
    
      spline_dfs_out_count <- lapply(seq_along(splines_list_out_count), function(i) {
        data.frame(
          weight_timestamp_numeric = one_hive_one_day$weight_timestamp_numeric,
          out_count_diff_spline = predict(splines_list_out_count[[i]], one_hive_one_day$weight_timestamp_numeric)$y, 
          spar = paste0("out_count Spline Fit spar=", spar_vals[i])
        ) %>% mutate(weight_timestamp = as.POSIXct(weight_timestamp_numeric, origin = "1970-01-01"))
      })
    
      combined_spline_df_out_count <- do.call(rbind, spline_dfs_out_count)
    
      # compare imputed data with actual data: out_count
      one_hive_one_day$out_count_diff_original <- one_hive_one_day$out_count - lag(one_hive_one_day$out_count)
      one_hive_one_day$out_count_diff_original[is.na(one_hive_one_day$out_count_diff_original)] <- 0
    
      plot4 <- ggplot(one_hive_one_day, aes(x = weight_timestamp)) +
        geom_point(aes(y = out_count_diff_imputed, color = "Imputed Data"), na.rm = TRUE) +
        geom_point(aes(y = out_count_diff_original, color = "Original Data"), na.rm = TRUE) +
        geom_line(data = combined_spline_df_out_count, aes(x = weight_timestamp, y = out_count_diff_spline, color = spar), size = 1) +
        geom_vline(xintercept = end_time, linetype = "dashed", color = "red") +
        annotate("text", x = end_time, y = max(one_hive_one_day$out_count_diff, na.rm = TRUE), label = "End Missing Data", color = "red", angle = 90, vjust = -0.5, hjust = 1) +
        labs(title = paste0("Hive ", one_hive, ": out_count Imputations For Timespan Ending ", end_time), x = "Timestamp", y = "in_count_diff") +
        scale_color_brewer(palette = "Set1") +
        theme_minimal() +
        theme(legend.title = element_blank())
      
      print(plot4)
      
      print(paste0("timestamps_remaining is now ", timestamps_remaining))
    } else {
      print(paste0("Skipping the remainder of plots for hive", one_hive, "- not enough unique timestamp values\n"))
      # next
    }
  }
}
  
```
