==> 3a-plotting-centroid.py <==
import pandas as pd
import matplotlib.pyplot as plt

audio_data = pd.read_csv("C:/ncf-graduate-school/internship-USDA/almond-pollination/data/sensor_sound_merged.csv")

audio_data_unscaled = pd.DataFrame(audio_data, columns=['timestamp', 'pcb_temperature_one', 'humidity', 'centroid', 'C1_10', 'C11_20', 'C21_30', 'C31_40', 'C41_50', 'C51_60', 'C61_70', 'C71_80', 'C81_90', 'C91_100', 'C101_110', 'C111_120', 'C121_130', 'C131_140', 'C141_150', 'C151_160', 'C161_170', 'C171_180', 'C181_190', 'C191_200', 'C201_210', 'C211_220', 'C221_230', 'C231_240', 'C241_250', 'C251_260', 'C261_270', 'C271_280', 'C281_290', 'C291_300', 'C301_310', 'C311_320', 'C321_330', 'C331_340', 'C341_350', 'C351_360', 'C361_370', 'C371_380', 'C381_390', 'C391_400', 'C401_410', 'C411_420', 'C421_430', 'C431_440', 'C441_450', 'C451_460', 'C461_470', 'C471_480', 'C481_490', 'C491_500', 'C501_510', 'C511_520', 'C521_530', 'C531_540', 'C541_550', 'C551_560', 'C561_570', 'C571_580', 'C581_590', 'C591_600', 'C601_610', 'C611_620', 'C621_630', 'C631_640', 'C641_650', 'C651_660', 'C661_670', 'C671_680', 'C681_690', 'C691_700', 'C701_710', 'C711_720', 'C721_730', 'C731_740', 'C741_750', 'C751_760', 'C761_770', 'C771_780', 'C781_790', 'C791_800', 'C801_810', 'C811_820', 'C821_830', 'C831_840', 'C841_850', 'C851_860', 'C861_870', 'C871_880', 'C881_890', 'C891_900', 'C901_910', 'C911_920', 'C921_930', 'C931_940', 'C941_950', 'C951_960', 'C961_970', 'C971_980', 'C981_990', 'C991_1000', 'C1001_1010', 'C1011_1020', 'C1021_1030', 'C1031_1040', 'C1041_1050', 'C1051_1060', 'C1061_1070', 'C1071_1080', 'C1081_1090', 'C1091_1100', 'C1101_1110', 'C1111_1120', 'C1121_1130', 'C1131_1140', 'C1141_1150', 'C1151_1160', 'C1161_1170', 'C1171_1180', 'C1181_1190', 'C1191_1200', 'C1201_1210', 'C1211_1220', 'C1221_1230', 'C1231_1240', 'C1241_1250', 'C1251_1260', 'C1261_1270', 'C1271_1280', 'C1281_1290', 'C1291_1300', 'C1301_1310', 'C1311_1320', 'C1321_1330', 'C1331_1340', 'C1341_1350', 'C1351_1360', 'C1361_1370', 'C1371_1380', 'C1381_1390', 'C1391_1400', 'C1401_1410', 'C1411_1420', 'C1421_1430', 'C1431_1440', 'C1441_1450', 'C1451_1460', 'C1461_1470', 'C1471_1480', 'C1481_1490', 'C1491_1500', 'C1501_1510', 'C1511_1520', 'C1521_1530', 'C1531_1540', 'C1541_1550', 'C1551_1560', 'C1561_1570', 'C1571_1580', 'C1581_1590', 'C1591_1600', 'C1601_1610', 'C1611_1620', 'C1621_1630', 'C1631_1640', 'C1641_1650', 'C1651_1660', 'C1661_1670', 'C1671_1680', 'C1681_1690', 'C1691_1700', 'C1701_1710', 'C1711_1720', 'C1721_1730', 'C1731_1740', 'C1741_1750', 'C1751_1760', 'C1761_1770', 'C1771_1780', 'C1781_1790', 'C1791_1800', 'C1801_1810', 'C1811_1820', 'C1821_1830', 'C1831_1840', 'C1841_1850', 'C1851_1860', 'C1861_1870', 'C1871_1880', 'C1881_1890', 'C1891_1900', 'C1901_1910', 'C1911_1920', 'C1921_1930', 'C1931_1940', 'C1941_1950', 'C1951_1960', 'C1961_1970', 'C1971_1980', 'C1981_1990', 'C1991_2000', 'C2001_2010', 'C2011_2020', 'C2021_2030', 'C2031_2040', 'C2041_2050', 'C2051_2060', 'C2061_2070', 'C2071_2080', 'C2081_2090', 'C2091_2100', 'C2101_2110', 'C2111_2120', 'C2121_2130', 'C2131_2140', 'C2141_2150', 'C2151_2160', 'C2161_2170', 'C2171_2180', 'C2181_2190', 'C2191_2200', 'C2201_2210', 'C2211_2220', 'C2221_2230', 'C2231_2240', 'C2241_2250', 'C2251_2260', 'C2261_2270', 'C2271_2280', 'C2281_2290', 'C2291_2300', 'C2301_2310', 'C2311_2320', 'C2321_2330', 'C2331_2340', 'C2341_2350', 'C2351_2360', 'C2361_2370', 'C2371_2380', 'C2381_2390', 'C2391_2400', 'C2401_2410', 'C2411_2420', 'C2421_2430', 'C2431_2440', 'C2441_2450', 'C2451_2460', 'C2461_2470', 'C2471_2480', 'C2481_2490', 'C2491_2500', 'C2501_2510', 'C2511_2520', 'C2521_2530', 'C2531_2540', 'C2541_2550', 'C2551_2560', 'C2561_2570', 'C2571_2580', 'C2581_2590', 'C2591_2600', 'C2601_2610', 'C2611_2620', 'C2621_2630', 'C2631_2640', 'C2641_2650', 'C2651_2660', 'C2661_2670', 'C2671_2680', 'C2681_2690', 'C2691_2700', 'C2701_2710', 'C2711_2720', 'C2721_2730', 'C2731_2740', 'C2741_2750', 'C2751_2760', 'C2761_2770', 'C2771_2780', 'C2781_2790', 'C2791_2800', 'C2801_2810', 'C2811_2820', 'C2821_2830', 'C2831_2840', 'C2841_2850', 'C2851_2860', 'C2861_2870', 'C2871_2880', 'C2881_2890', 'C2891_2900', 'C2901_2910', 'C2911_2920', 'C2921_2930', 'C2931_2940', 'C2941_2950', 'C2951_2960', 'C2961_2970', 'C2971_2980', 'C2981_2990', 'C2991_3000', 'C3001_3010', 'C3011_3020', 'C3021_3030', 'C3031_3040', 'C3041_3050', 'C3051_3060', 'C3061_3070', 'C3071_3080', 'C3081_3090', 'C3091_3100', 'C3101_3110', 'C3111_3120', 'C3121_3130', 'C3131_3140', 'C3141_3150', 'C3151_3160', 'C3161_3170', 'C3171_3180', 'C3181_3190', 'C3191_3200', 'C3201_3210', 'C3211_3220', 'C3221_3230', 'C3231_3240', 'C3241_3250', 'C3251_3260', 'C3261_3270', 'C3271_3280', 'C3281_3290', 'C3291_3300', 'C3301_3310', 'C3311_3320', 'C3321_3330', 'C3331_3340', 'C3341_3350', 'C3351_3360', 'C3361_3370', 'C3371_3380', 'C3381_3390', 'C3391_3400', 'C3401_3410', 'C3411_3420', 'C3421_3430', 'C3431_3440', 'C3441_3450', 'C3451_3460', 'C3461_3470', 'C3471_3480', 'C3481_3490', 'C3491_3500', 'C3501_3510', 'C3511_3520', 'C3521_3530', 'C3531_3540', 'C3541_3550', 'C3551_3560', 'C3561_3570', 'C3571_3580', 'C3581_3590', 'C3591_3600', 'C3601_3610', 'C3611_3620', 'C3621_3630', 'C3631_3640', 'C3641_3650', 'C3651_3660', 'C3661_3670', 'C3671_3680', 'C3681_3690', 'C3691_3700', 'C3701_3710', 'C3711_3720', 'C3721_3730', 'C3731_3740', 'C3741_3750', 'C3751_3760', 'C3761_3770', 'C3771_3780', 'C3781_3790', 'C3791_3800', 'C3801_3810', 'C3811_3820', 'C3821_3830', 'C3831_3840', 'C3841_3850', 'C3851_3860', 'C3861_3870', 'C3871_3880', 'C3881_3890', 'C3891_3900', 'C3901_3910', 'C3911_3920', 'C3921_3930', 'C3931_3940', 'C3941_3950', 'C3951_3960', 'C3961_3970', 'C3971_3980', 'C3981_3990', 'C3991_4000'])

audio_data_unscaled = audio_data_unscaled.iloc[0:40000, :]

audio_data_unscaled['timestamp'] = pd.to_datetime(audio_data_unscaled['timestamp'])
audio_data_unscaled['timestamp'] = audio_data_unscaled['timestamp'].astype('int64')

x = audio_data_unscaled['timestamp']
y = audio_data_unscaled['centroid']

# plot centroid vs timestamp
plt.figure(figsize=(10, 6))
plt.scatter(x, y, marker='o', color='b', label='Centroid vs Timestamp')
plt.xlabel('Timestamp')
plt.ylabel('Centroid')
plt.title('Centroid vs Timestamp')
plt.legend()
plt.grid(True)

x = audio_data_unscaled['pcb_temperature_one']
y = audio_data_unscaled['centroid']

# plot centroid vs temp 
plt.figure(figsize=(10, 6))
plt.scatter(x, y, marker='o', color='b', label='Centroid vs Temperature')
plt.xlabel('Temperature')
plt.ylabel('Centroid')
plt.xlim(left=0)
plt.title('Centroid vs Temperature')
plt.legend()
plt.grid(True)

x = audio_data_unscaled['humidity']
y = audio_data_unscaled['centroid']

# plot centroid vs humidity 
plt.figure(figsize=(10, 6))
plt.scatter(x, y, marker='o', color='b', label='Centroid vs Humidity')
plt.xlabel('Humidity')
plt.ylabel('Centroid')
plt.xlim(left=0)
plt.title('Centroid vs Humidity')
plt.legend()
plt.grid(True)

plt.show()
